+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 
 - All kafka components now detect unrecoverable connection issues and back off more aggressively. (@Jeffail)
+- The `redpanda_migrator_offsets` input now fetches record timestamps in parallel and discards consumer groups which point to truncated records. (@mihaitodor)
 
 ## 4.59.0 - 2025-06-27